// src/app/interfaces/location.ts
export interface Root2 {
  place_id: string
  osm_id: string
  osm_type: string
  licence: string
  lat: string
  lon: string
  boundingbox: string[]
  class: string
  type: string
  display_name: string
  display_place: string
  display_address: string
  address: Address
}

export interface Address {
  name: string
  house_number: string
  road: string
  neighbourhood?: string
  suburb?: string
  town?: string
  village?: string
  city?: string
  county?: string
  state: string
  postcode: string
  country: string
  country_code: string
}